+++ lightroom/photara.lrplugin/Info.lua
@@ -1,0 +1,25 @@
+return {
+    LrSdkVersion = 13.0,
+    LrSdkMinimumVersion = 6.0,
+
+    LrToolkitIdentifier = "com.orbyts.photara",
+    LrPluginName = "Photara",
+
+    LrLibraryMenuItems = {
+        {
+            title = "Validate Photara Connection",
+            file = "ValidateMain.lua",
+        },
+        {
+            title = "Apply Project to Selected Shoot",
+            file = "ApplyProjectMain.lua",
+        },
+    },
+
+    VERSION = {
+        major = 0,
+        minor = 0,
+        revision = 5,
+        build = 0,
+    },
+}